Add tools for usb wifi dealie
Just spent a long ass time trying to get my netgear wnda3100v2 usb wifi
adapter to work with ubuntu. There was a very specific version of
ndiswrapper and of the actual windows driver (which was hard to find) that
are required for it to work. I'm adding it here so that I don't have to
suffer through it again.